Move settings component into controls component
