+++ src/App.vue
@@ -4,51 +4,29 @@
   <Controls :controller='controller' @openSettings="openSettings" />
   <Quest title='123' :controller='controller' />
   <button class="open-settings" @click='$refs.Settings.openSettings()'>Open settings</button>
-  <Settings
-      ref='Settings'
-      :controller='controller'
-      :show="showSettings"
-      @openSettings="openSettings"
-      @closeSettings="closeSettings"
-  />
 </template>
 
 <script>
-import { ref, reactive } from 'vue'
+import { ref } from 'vue'
 import Title from './components/Title.vue'
 import Controls from './components/Controls.vue'
 import Quest from './components/Quest.vue'
 import Onetwothree from './classOnetwothree'
-import Settings from './components/Settings'
 
 export default {
   name: 'App',
   components: {
     Title,
     Controls,
-    Quest,
-    Settings
+    Quest
   },
   setup () {
-    const showSettings = ref(false)
-
-    const controller = reactive(new Onetwothree({
+    const controller = ref(new Onetwothree({
       shuffle: true
     }))
 
-    const openSettings = function () {
-      showSettings.value = true
-    }
-
-    const closeSettings = function () {
-      showSettings.value = false
-    }
-
     return {
-      controller,
-      openSettings,
-      closeSettings,
-      showSettings
+      controller
     }
   }
 }
